requirements.txt
llama-index
google-genai
qdrant-client
llama-index-llms-google_genai
llama-index-vector-stores-qdrant
llama-index-embeddings-google_genai
llama-index-readers-file
llama-index-core
fastapi
uvicorn
pydantic
